+++ hypha/apply/activity/adapters/__init__.py
@@ -1,0 +1,13 @@
+from .activity_feed import ActivityAdapter
+from .base import AdapterBase
+from .django_messages import DjangoMessagesAdapter
+from .emails import EmailAdapter
+from .slack import SlackAdapter
+
+__all__ = [
+    AdapterBase,
+    ActivityAdapter,
+    DjangoMessagesAdapter,
+    EmailAdapter,
+    SlackAdapter,
+]
